% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{create_configuration_set}
\alias{create_configuration_set}
\title{Creates a configuration set}
\usage{
create_configuration_set(ConfigurationSet)
}
\arguments{
\item{ConfigurationSet}{[required] A data structure that contains the name of the configuration set.}
}
\description{
Creates a configuration set.
}
\details{
Configuration sets enable you to publish email sending events. For information about using configuration sets, see the \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/monitor-sending-activity.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\section{Accepted Parameters}{
\preformatted{create_configuration_set(
  ConfigurationSet = list(
    Name = "string"
  )
)
}
}

