% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{create_configuration_set_event_destination}
\alias{create_configuration_set_event_destination}
\title{Creates a configuration set event destination}
\usage{
create_configuration_set_event_destination(ConfigurationSetName,
  EventDestination)
}
\arguments{
\item{ConfigurationSetName}{The name of the configuration set that the event destination should be associated with.}

\item{EventDestination}{An object that describes the AWS service that email sending event information will be published to.}
}
\description{
Creates a configuration set event destination.
}
\details{
When you create or update an event destination, you must provide one, and only one, destination. The destination can be CloudWatch, Amazon Kinesis Firehose, or Amazon Simple Notification Service (Amazon SNS).

An event destination is the AWS service to which Amazon SES publishes the email sending events associated with a configuration set. For information about using configuration sets, see the \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/monitor-sending-activity.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\examples{

}
