% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{create_receipt_rule}
\alias{create_receipt_rule}
\title{Creates a receipt rule}
\usage{
create_receipt_rule(RuleSetName, Rule, After = NULL)
}
\arguments{
\item{RuleSetName}{The name of the rule set that the receipt rule will be added to.}

\item{Rule}{A data structure that contains the specified rule's name, actions, recipients, domains, enabled status, scan status, and TLS policy.}

\item{After}{The name of an existing rule after which the new rule will be placed. If this parameter is null, the new rule will be inserted at the beginning of the rule list.}
}
\description{
Creates a receipt rule.
}
\details{
For information about setting up receipt rules, see the \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/receiving-email-receipt-rules.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\examples{
# The following example creates a new receipt rule:
create_receipt_rule(
  After = "",
  Rule = list(
    Actions = list(
      list(
        S3Action = list(
          BucketName = "MyBucket",
          ObjectKeyPrefix = "email"
        )
      )
    ),
    Enabled = TRUE,
    Name = "MyRule",
    ScanEnabled = TRUE,
    TlsPolicy = "Optional"
  ),
  RuleSetName = "MyRuleSet"
)

}
