% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{create_template}
\alias{create_template}
\title{Creates an email template}
\usage{
create_template(
  Template = list(
    TemplateName = "string",
    SubjectPart = "string",
    TextPart = "string",
    HtmlPart = "string"
  )
)
}
\arguments{
\item{Template}{[required] The content of the email, composed of a subject line, an HTML part, and a text-only part.}
}
\description{
Creates an email template. Email templates enable you to send personalized email to one or more destinations in a single API operation. For more information, see the \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/send-personalized-email-api.html}{Amazon SES Developer Guide}.
}
\details{
You can execute this operation no more than once per second.
}
\examples{

}
