% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{delete_configuration_set}
\alias{delete_configuration_set}
\title{Deletes a configuration set}
\usage{
delete_configuration_set(ConfigurationSetName)
}
\arguments{
\item{ConfigurationSetName}{The name of the configuration set to delete.}
}
\description{
Deletes a configuration set. Configuration sets enable you to publish email sending events. For information about using configuration sets, see the \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/monitor-sending-activity.html}{Amazon SES Developer Guide}.
}
\details{
You can execute this operation no more than once per second.
}
\examples{

}
