% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{delete_custom_verification_email_template}
\alias{delete_custom_verification_email_template}
\title{Deletes an existing custom verification email template}
\usage{
delete_custom_verification_email_template(TemplateName)
}
\arguments{
\item{TemplateName}{The name of the custom verification email template that you want to delete.}
}
\description{
Deletes an existing custom verification email template.
}
\details{
For more information about custom verification email templates, see \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/custom-verification-emails.html}{Using Custom Verification Email Templates} in the \emph{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\examples{

}
