% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{delete_identity}
\alias{delete_identity}
\title{Deletes the specified identity (an email address or a domain) from the list of verified identities}
\usage{
delete_identity(
  Identity = "string"
)
}
\arguments{
\item{Identity}{[required] The identity to be removed from the list of identities for the AWS Account.}
}
\description{
Deletes the specified identity (an email address or a domain) from the list of verified identities.
}
\details{
You can execute this operation no more than once per second.
}
\examples{
# The following example deletes an identity from the list of identities
# that have been submitted for verification with Amazon SES:
delete_identity(
  Identity = "user@example.com"
)

}
