% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{delete_receipt_filter}
\alias{delete_receipt_filter}
\title{Deletes the specified IP address filter}
\usage{
delete_receipt_filter(FilterName)
}
\arguments{
\item{FilterName}{The name of the IP address filter to delete.}
}
\description{
Deletes the specified IP address filter.
}
\details{
For information about managing IP address filters, see the \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/receiving-email-managing-ip-filters.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\examples{
# The following example deletes an IP address filter:
delete_receipt_filter(
  FilterName = "MyFilter"
)

}
