% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{delete_template}
\alias{delete_template}
\title{Deletes an email template}
\usage{
delete_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the template to be deleted.}
}
\description{
Deletes an email template.
}
\details{
You can execute this operation no more than once per second.
}
\section{Accepted Parameters}{
\preformatted{delete_template(
  TemplateName = "string"
)
}
}

