% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{get_custom_verification_email_template}
\alias{get_custom_verification_email_template}
\title{Returns the custom email verification template for the template name you specify}
\usage{
get_custom_verification_email_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the custom verification email template that you want to retrieve.}
}
\description{
Returns the custom email verification template for the template name you specify.
}
\details{
For more information about custom verification email templates, see \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/custom-verification-emails.html}{Using Custom Verification Email Templates} in the \emph{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\section{Accepted Parameters}{
\preformatted{get_custom_verification_email_template(
  TemplateName = "string"
)
}
}

