% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{get_identity_mail_from_domain_attributes}
\alias{get_identity_mail_from_domain_attributes}
\title{Returns the custom MAIL FROM attributes for a list of identities (email addresses : domains)}
\usage{
get_identity_mail_from_domain_attributes(Identities)
}
\arguments{
\item{Identities}{A list of one or more identities.}
}
\description{
Returns the custom MAIL FROM attributes for a list of identities (email addresses : domains).
}
\details{
This operation is throttled at one request per second and can only get custom MAIL FROM attributes for up to 100 identities at a time.
}
\examples{
# The following example returns the custom MAIL FROM attributes for an
# identity:
get_identity_mail_from_domain_attributes(
  Identities = list(
    "example.com"
  )
)

}
