% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{get_send_quota}
\alias{get_send_quota}
\title{Provides the sending limits for the Amazon SES account}
\usage{
get_send_quota()
}
\description{
Provides the sending limits for the Amazon SES account.
}
\details{
You can execute this operation no more than once per second.
}
\examples{
# The following example returns the Amazon SES sending limits for an AWS
# account:
get_send_quota()

}
