% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{get_template}
\alias{get_template}
\title{Displays the template object (which includes the Subject line, HTML part and text part) for the template you specify}
\usage{
get_template(TemplateName)
}
\arguments{
\item{TemplateName}{[required] The name of the template you want to retrieve.}
}
\description{
Displays the template object (which includes the Subject line, HTML part and text part) for the template you specify.
}
\details{
You can execute this operation no more than once per second.
}
\section{Accepted Parameters}{
\preformatted{get_template(
  TemplateName = "string"
)
}
}

