% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{list_identities}
\alias{list_identities}
\title{Returns a list containing all of the identities (email addresses and domains) for your AWS account in the current AWS Region, regardless of verification status}
\usage{
list_identities(
  IdentityType = "EmailAddress"|"Domain",
  NextToken = "string",
  MaxItems = 123
)
}
\arguments{
\item{IdentityType}{The type of the identities to list. Possible values are "EmailAddress" and "Domain". If this parameter is omitted, then all identities will be listed.}

\item{NextToken}{The token to use for pagination.}

\item{MaxItems}{The maximum number of identities per page. Possible values are 1-1000 inclusive.}
}
\description{
Returns a list containing all of the identities (email addresses and domains) for your AWS account in the current AWS Region, regardless of verification status.
}
\details{
You can execute this operation no more than once per second.
}
\examples{
# The following example lists the email address identities that have been
# submitted for verification with Amazon SES:
list_identities(
  IdentityType = "EmailAddress",
  MaxItems = 123L,
  NextToken = ""
)

}
