% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{list_templates}
\alias{list_templates}
\title{Lists the email templates present in your Amazon SES account in the current AWS Region}
\usage{
list_templates(NextToken = NULL, MaxItems = NULL)
}
\arguments{
\item{NextToken}{A token returned from a previous call to \code{ListTemplates} to indicate the position in the list of email templates.}

\item{MaxItems}{The maximum number of templates to return. This value must be at least 1 and less than or equal to 10. If you do not specify a value, or if you specify a value less than 1 or greater than 10, the operation will return up to 10 results.}
}
\description{
Lists the email templates present in your Amazon SES account in the current AWS Region.
}
\details{
You can execute this operation no more than once per second.
}
\examples{

}
