% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_package.R
\docType{package}
\name{paws.ses-package}
\alias{paws.ses}
\alias{paws.ses-package}
\title{paws.ses: Amazon Simple Email Service}
\description{
This document contains reference information for
the [Amazon Simple Email Service](https://aws.amazon.com/ses/) (Amazon
SES) API, version 2010-12-01. This document is best used in
conjunction with the [Amazon SES Developer
Guide](http://docs.aws.amazon.com/ses/latest/DeveloperGuide/Welcome.html).
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
