% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{set_receipt_rule_position}
\alias{set_receipt_rule_position}
\title{Sets the position of the specified receipt rule in the receipt rule set}
\usage{
set_receipt_rule_position(RuleSetName, RuleName, After = NULL)
}
\arguments{
\item{RuleSetName}{The name of the receipt rule set that contains the receipt rule to reposition.}

\item{RuleName}{The name of the receipt rule to reposition.}

\item{After}{The name of the receipt rule after which to place the specified receipt rule.}
}
\description{
Sets the position of the specified receipt rule in the receipt rule set.
}
\details{
For information about managing receipt rules, see the \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/receiving-email-managing-receipt-rules.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\examples{
# The following example sets the position of a receipt rule in a receipt
# rule set:
set_receipt_rule_position(
  After = "PutRuleAfterThisRule",
  RuleName = "RuleToReposition",
  RuleSetName = "MyRuleSet"
)

}
