% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{update_receipt_rule}
\alias{update_receipt_rule}
\title{Updates a receipt rule}
\usage{
update_receipt_rule(RuleSetName, Rule)
}
\arguments{
\item{RuleSetName}{The name of the receipt rule set that the receipt rule belongs to.}

\item{Rule}{A data structure that contains the updated receipt rule information.}
}
\description{
Updates a receipt rule.
}
\details{
For information about managing receipt rules, see the \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/receiving-email-managing-receipt-rules.html}{Amazon SES Developer Guide}.

You can execute this operation no more than once per second.
}
\examples{
# The following example updates a receipt rule to use an Amazon S3 action:
update_receipt_rule(
  Rule = list(
    Actions = list(
      list(
        S3Action = list(
          BucketName = "MyBucket",
          ObjectKeyPrefix = "email"
        )
      )
    ),
    Enabled = TRUE,
    Name = "MyRule",
    ScanEnabled = TRUE,
    TlsPolicy = "Optional"
  ),
  RuleSetName = "MyRuleSet"
)

}
