% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{update_template}
\alias{update_template}
\title{Updates an email template}
\usage{
update_template(
  Template = list(
    TemplateName = "string",
    SubjectPart = "string",
    TextPart = "string",
    HtmlPart = "string"
  )
)
}
\arguments{
\item{Template}{[required]}
}
\description{
Updates an email template. Email templates enable you to send personalized email to one or more destinations in a single API operation. For more information, see the \href{http://docs.aws.amazon.com/ses/latest/DeveloperGuide/send-personalized-email-api.html}{Amazon SES Developer Guide}.
}
\details{
You can execute this operation no more than once per second.
}
\examples{

}
