% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{verify_email_identity}
\alias{verify_email_identity}
\title{Adds an email address to the list of identities for your Amazon SES account in the current AWS region and attempts to verify it}
\usage{
verify_email_identity(
  EmailAddress = "string"
)
}
\arguments{
\item{EmailAddress}{[required] The email address to be verified.}
}
\description{
Adds an email address to the list of identities for your Amazon SES account in the current AWS region and attempts to verify it. As a result of executing this operation, a verification email is sent to the specified address.
}
\details{
You can execute this operation no more than once per second.
}
\examples{
# The following example starts the email address verification process with
# Amazon SES:
verify_email_identity(
  EmailAddress = "user@example.com"
)

}
