% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{delete_verified_email_address}
\alias{delete_verified_email_address}
\title{Deprecated}
\usage{
delete_verified_email_address(EmailAddress)
}
\arguments{
\item{EmailAddress}{[required] An email address to be removed from the list of verified addresses.}
}
\description{
Deprecated. Use the \code{DeleteIdentity} operation to delete email addresses and domains.
}
\section{Accepted Parameters}{
\preformatted{delete_verified_email_address(
  EmailAddress = "string"
)
}
}

\examples{
# The following example deletes an email address from the list of
# identities that have been submitted for verification with Amazon SES:
\donttest{delete_verified_email_address(
  EmailAddress = "user@example.com"
)}

}
