% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ses_operations.R
\name{verify_email_address}
\alias{verify_email_address}
\title{Deprecated}
\usage{
verify_email_address(EmailAddress)
}
\arguments{
\item{EmailAddress}{[required] The email address to be verified.}
}
\description{
Deprecated. Use the \code{VerifyEmailIdentity} operation to verify a new email address.
}
\section{Accepted Parameters}{
\preformatted{verify_email_address(
  EmailAddress = "string"
)
}
}

\examples{
# The following example starts the email address verification process with
# Amazon SES:
\donttest{verify_email_address(
  EmailAddress = "user@example.com"
)}

}
